requirements.txt
python-dotenv
google-genai
matplotlib
pillow
openpyxl
